# docs/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: GreenRoot API
  version: 0.1.0
  description: Modular Go chi backend for the GreenRoot platform.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/auth/send-otp:
    post:
      tags: [Auth]
      summary: Send OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOTPRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendOTPResponse"
        "400":
          description: Invalid request
  /api/v1/auth/verify-otp:
    post:
      tags: [Auth]
      summary: Verify OTP and login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOTPRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid OTP
  /api/v1/auth/refresh-token:
    post:
      tags: [Auth]
      summary: Refresh JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid refresh token
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout
      security:
        - BearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
      responses:
        "200":
          description: Logged out
        "401":
          description: Invalid refresh token
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Current user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Invalid access token
...
  /api/v1/inventory:
    get:
      tags: [Inventory]
      summary: List inventory
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
          description: Search by nursery, scientific name, common name, or size.
        - name: nursery_id
          in: query
          schema:
            type: integer
        - name: plant_id
          in: query
          schema:
            type: integer
        - name: size_id
          in: query
          schema:
            type: integer
        - name: inventory_status
          in: query
          schema:
            type: string
            enum: [AVAILABLE, LOW_STOCK, OUT_OF_STOCK, RESERVED, DISCONTINUED]
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [id, nursery_name, scientific_name, common_name, size_code, available_quantity, inventory_status, last_updated_at, created_at]
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: Inventory list
    post:
      tags: [Inventory]
      summary: Create inventory item
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryRequest"
      responses:
        "201":
          description: Inventory item created
  /api/v1/inventory/{id}:
    get:
      tags: [Inventory]
      summary: Get inventory item
      parameters:
        - $ref: "#/components/parameters/InventoryID"
      responses:
        "200":
          description: Inventory item
    put:
      tags: [Inventory]
      summary: Update inventory item
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/InventoryID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryRequest"
      responses:
        "200":
          description: Inventory item updated
    delete:
      tags: [Inventory]
      summary: Delete inventory item
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/InventoryID"
      responses:
        "200":
          description: Inventory item deleted
...
components:
  parameters:
...
    InventoryID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SendOTPRequest:
      type: object
      required: [mobile]
      properties:
        mobile:
          type: string
          example: "9000000002"
    VerifyOTPRequest:
      type: object
      required: [mobile, otp]
      properties:
        mobile:
          type: string
          example: "9000000002"
        otp:
          type: string
          example: "123456"
        device_type:
          type: string
          example: "MOBILE"
        os_name:
          type: string
          example: "ANDROID"
        app_version:
          type: string
          example: "1.0.0"
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    LogoutRequest:
      type: object
      properties:
        refresh_token:
          type: string
    SendOTPResponse:
      type: object
      properties:
        message:
          type: string
        mock_otp:
          type: string
          example: "123456"
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        user:
          $ref: "#/components/schemas/User"
    MeResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_code:
          type: string
          example: USR-000001
        first_name:
          type: string
        mobile:
          type: string
        mobile_verified:
          type: boolean
        email_verified:
          type: boolean
        status:
          type: string
        roles:
          type: array
          items:
            type: string
...
    InventoryRequest:
      type: object
      required: [nursery_id, plant_id, size_id, available_quantity]
      properties:
        nursery_id:
          type: integer
          format: int64
        plant_id:
          type: integer
          format: int64
        size_id:
          type: integer
        available_quantity:
          type: integer
          minimum: 0
        inventory_status:
          type: string
          enum: [AVAILABLE, LOW_STOCK, OUT_OF_STOCK, RESERVED, DISCONTINUED]
...